pyproject.toml
[tool.poetry]
name = "medical-rag"
version = "0.1.0"
description = ""
authors = ["ArsalanAyaz <henrykoff7788@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
langchain = "^0.3.0"
faiss-cpu = "^1.8.0.post1"
langchain-google-genai = "^2.0.0"
langchain-tools = "^0.1.34"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.22"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.147.0"
langchain-community = "^0.3.0"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
